+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 7.6.10 (01/18/2021 - )
 
 * api: for max process time, webservice will consider both http client timeout and remaining process time of current action
+* monitor: support multiple alert channels
 
 ### 7.6.9 (01/04/2021 - 01/15/2021)